Fix commit_transaction_block_checkpoint test.
The fault injection points used in the test didn't exist anymore. Add a new
injection point in RecordTransactionCommit(), just before writing the commit
WAL record, and use that in the test.

Remove a bunch of fault injection IDs that are no longer used. (They are
still referenced in some TINC tests, but the injection points don't exist
anymore, so those tests will need to be rewritten if we want to keep them.)