Remove remnants of persistent tables.
They were not kept up-to-date anymore anyway. Remove the actual tables.

There are still a few references to these tables in the management tools.
AFAICS they're in tests, and I was hesitent to remove them just yet, in
case we're going to use the existing tests as a guide when writing new
tests.