Remove checks related to persistent tables from gpcheckcat.
Because persistent tables are no more.

NOTE: It would still be nice to check for consistency between pg_class
and files on disk, to check that there are no extra data files, and no
data files missing that have a pg_class entry. Same with AO seg files,
I suppose. But that's a significantly different query than what we have
here.