WAL-log creation of empty AO segfiles.
An empty segfile is mostly treated the same as a missing segfile, but for
the sake of gp_replica_check, WAL-log the creation of an empty segfile
anyway, so that there is no inconsistency between master and mirror, such
that an empty segfile exists on master, but it's missing entirely in the
mirror. (I'm not entirely sure if there is non-testing code that requires
that, too, so better safe than sorry).

This should fix the warnings like this:

WARNING:  Unable to open file /tmp/build/e18b2f02/gpdb_src/gpAux/gpdemo/datadirs/dbfast_mirror2/demoDataDir1/base/16384/61117.1152

from gp_replica_check. (There are other failures still.)