Work around the fact that nothing might a flush WAL, in gp_replica_check.
gp_replica_check would often get stuck, waiting for the standby to apply all
the WAL it was sent. However, there is nothing to force a WAL flush in the
master. Usually, the last record in a transaction is a transaction commit,
which is flushed, and many other things cause a WAL flush too, but when
running the regression suite, often the last WAL record is a WAL-logged
hint bit update, just after a checkpoint.

To work around that, if the standby doesn't catch up in 20 seconds, issue
a CHECKPOINT in the master, to force a WAL flush. Something more
lightweight could be used to flush the WAL, but gp_replica_check needs the
data on disk to be up to date, so a checkpoint seems like a good idea.
In fact, perhaps we should always issue a CHECKPOINT, even before the first
attempt. Currently the python script does that, but now it seems redundant..