Restore logic to skip databases cannot connect for oldest database.
GPDB skips databases that cannot be connected to in computing the oldest
database in vac_truncate_clog(). Make write_database_file() same, which was
reverted to upstream version. This helps to get the storage tests green for now.

Later can figure out and uniformly remove this code from vac_truncate_clog() and
write_database_file() if better solution is found to original issue for which
this check was added.