gpstop -u should not specifically check for "No such process"
If postmaster.pid file is present, reload will get error as "No such
process". But if postmaster.pid is not present then error returned back is
"pg_ctl: PID file "......../postmaster.pid" does not exist". So, its better not
to check for any particular error message but report segmnes failed to be
reloaded.