Embed dbid in tablespace path, to avoid clash between servers on same host.
This is a backport upstream commit 228170410d, from PostgreSQL 9.0, which
added the server version number in the path. But in GPDB, also include the
gp_dbid in the path. This makes it possible to use the same tablespace path
on multiple servers running on the same host, without clashing.

Also includes cherry-pick of the small upstream cleanup commits 5c82ccb1dd
a6f56efc35, and c282b36dd2.

Re-enable upstream 'tablespace' regression test. It works now, even when
all the nodes are running on same host.