Bump up max_wal_senders, to fix src/test/walrep/ tests.
The test uses as a replication connection, so it didn't work with
max_wal_senders=1 if the QD node has a standby. And it seems like a good
idea to have some headroom, anyway. At some point, we ought to make this a
true GUC like in the upstream, but this will do for now.