Move walreceiver walrep test to run on mirrorless cluster
The walreceiver test establishes a walrep connection to the master's
walsender. With the recent change to start standby master by default
on gpdemo cluster, the mock walreceiver will be unable to create a
connection to the walsender since max_wal_senders is hardcoded to
1. Moving this to run only on mirrorless cluster makes more sense
since the assumption is to have an available walrep connection.