Disable gp_replica_check for sequences.
There are some "known" differences that are currently not masked out.
I think the 'last_value' field is WAL-logged to a higher value than what's
written to disk in the master.

This was not visible before, when the demo cluster didn't have a standby
for the QD node. In QE nodes, sequences are not used.