Add expected start state to commit_blocking test
Many times the `commit_blocking` test fails, its not clear if the
problem happened during the test, or cluster was not in right state
before we started. Hence adding expected state for the test at start
which would help to easily know if this test failed or is just
side-effect for previous test failure.