Make logic to wait for CHECKPOINT to be replayed more robust.
The code used to wait for sent_location == replay_location, but that had
similar problems that we had with the similar code in gp_replica_check
earlier.

Firstly, there is no guarantee that a checkpoint had been sent to the
mirror, by the time that we check for it. So even if sent_location ==
replay_location, it was possible that the checkpoint hadn't even been sent
to the mirror. That was highly unlikely in practice, though.

Secondly, there are some corner cases where the primary would send half
of a WAL record, up to a page boundary, but not the rest of the record,
until something nudges it. I think that happens if the WAL record is
not part of a transaction, so that there is no commit record to force it
to be flushed. In that case, replay_location would not catch up to
sent_location, because sent_location would point in the middle of a WAL
record, and the standby can't replay half of a WAL record. That situation
would persist until there's some new WAL activity to force it to be flushed,
which could take longer than the timeout used in the test (possibly
forever).

To fix, memorize the WAL posititions on each segment, after issuing the
CHECKPOINT, and wait until replay_location catches up to that position.
That more accurately reflects what we are trying to accomplish: we want
to wait until the checkpoint has been replayed, not necessarily that the
standby is completely in sync with the primary. (gp_replica_check needed
more strict logic, it wants the standby to really catch up to the primary,
not just up to the latest checkpoint.)

This hopefully fixes the occasional failures of the missing_xlog test we've
been seeing in the pipeline.