Don't allow CREATE/DROP TABLESPACE in transaction. Like in upstream.
We had allowed it in GPDB, when we had all the fancy persistent table
and filespace mechanisms to track the created-but-not-yet-committed
tablespace creation. The upstream mechanism is more crude, and can't
handle that gracefully. For example, if you abort the CREATE TABLESPACE
transaction, the directory structure would be left behind. Or worse,
when you DROP TABLESPACE, the underlying directory is deleted immediately,
even if the transaction aborts. Hence, put back the upstream limitation.
That's been acceptable for Postgres, and is surely acceptable for GPDB
as well.